refactor: Answers 일급컬랙션 적용
