+++ src/main/java/qna/domain/Answers.java
@@ -1,20 +1,36 @@
 package qna.domain;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+import javax.persistence.OrderBy;
+import org.hibernate.annotations.Where;
 import qna.CannotDeleteException;
 
 @Embeddable
 public class Answers {
 
-    private List<Answer> answers;
+    @OneToMany(mappedBy = "question", cascade = CascadeType.ALL)
+    @Where(clause = "deleted = false")
+    @OrderBy("id ASC")
+    private List<Answer> answers = new ArrayList<>();
 
     protected Answers() {
     }
 
     public Answers(List<Answer> answers) {
         this.answers = answers;
+    }
+
+    public List<Answer> getAnswers() {
+        return answers;
+    }
+
+    public void add(Answer answer) {
+        answers.add(answer);
     }
 
     public List<DeleteHistory> deleteAll(User writer) throws CannotDeleteException {
